Changed param type in Time Format
Former-commit-id: e06de6a90f5b0c04b23a89e9897e9c0c859c9915

+++ src/in/co/praveenkumar/mdroid/helper/TimeFormat.java
@@ -24,7 +24,7 @@ public class TimeFormat {
 	 *            in seconds
 	 * @return section title
 	 */
-	public static String getSection(long time) {
+	public static String getSection(int time) {
 		Calendar c = Calendar.getInstance();
 		int nowHour = c.get(Calendar.HOUR_OF_DAY);
 		int nowDay = c.get(Calendar.DATE);
@@ -85,7 +85,7 @@ public class TimeFormat {
 	 *            in seconds
 	 * @return formatted time
 	 */
-	public static String getMinimalTime(long time) {
+	public static String getMinimalTime(int time) {
 		Calendar c = Calendar.getInstance();
 		c.setTimeInMillis(time);
 		int hour = c.get(Calendar.HOUR_OF_DAY);
